Moved sync to jobs
